fix(auth): Set emulator credentials when using the Auth emulator
