+++ src/main/java/com/google/firebase/database/FirebaseDatabase.java
@@ -35,6 +35,7 @@ import com.google.firebase.database.util.EmulatorHelper;
 import com.google.firebase.database.utilities.ParsedUrl;
 import com.google.firebase.database.utilities.Utilities;
 import com.google.firebase.database.utilities.Validation;
+import com.google.firebase.internal.EmulatorCredentials;
 import com.google.firebase.internal.FirebaseService;
 
 import com.google.firebase.internal.SdkUtils;
@@ -404,28 +405,6 @@ public class FirebaseDatabase {
     @Override
     public void destroy() {
       instance.destroy();
-    }
-  }
-
-  private static class EmulatorCredentials extends GoogleCredentials {
-
-    EmulatorCredentials() {
-      super(newToken());
-    }
-
-    private static AccessToken newToken() {
-      return new AccessToken("owner",
-          new Date(System.currentTimeMillis() + TimeUnit.HOURS.toMillis(1)));
-    }
-
-    @Override
-    public AccessToken refreshAccessToken() {
-      return newToken();
-    }
-
-    @Override
-    public Map<String, List<String>> getRequestMetadata() throws IOException {
-      return ImmutableMap.of();
     }
   }
 }
